Improved Bender set implementation of array contents redistribution.
The old implementation would not distribute gaps uniformly over all windows,
which is problematic when insertions are repeatedly made in the same window
(under some additional circumstances), because rebalancing of higher-level
windows would not generate gaps in some of its child windows (which would then
still be overflowing).

Unfortunately, the new implementation appears to be slightly (around 5%)
slower when elements are inserted on random locations. This can probably be
resolved by a faster computation of where gaps are supposed to go.